fix: wrap useSearchParams in Suspense boundary
